Update extgcd.cpp
int -> long long

+++ extgcd.cpp
@@ -1,5 +1,5 @@
 long long extgcd(long long a, long long b, long long& x, long long& y){ //ax + by = gcd(a, b)
-    int d(a);
+    long long d(a);
     if(b){
         d = extgcd(b, a%b, y, x);
         y -= (a/b) * x;
